requirements.txt
anyio==3.5.0
fastapi==0.72.0
idna==3.3
loguru==0.5.3
pydantic==1.9.0
python-dotenv==0.19.2
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
